api/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-dotenv
alembic
pytest
pytest-asyncio
httpx
aiosqlite